Put back string interns since there is still code that uses them, would cause false negatives in equals

+++ framework/src/main/groovy/org/moqui/impl/entity/condition/FieldValueCondition.groovy
@@ -30,7 +30,7 @@ class FieldValueCondition extends EntityConditionImplBase {
     protected final EntityCondition.ComparisonOperator operator
     protected Object value
     protected boolean ignoreCase = false
-    protected Integer curHashCode = null
+    protected Integer curHashCode = (Integer) null
     protected static final Class thisClass = FieldValueCondition.class
 
     FieldValueCondition(EntityConditionFactoryImpl ecFactoryImpl,
